+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Mock assert postBody
   * Logs parse js objects correctly in console module
   * Puppeteer updated to 1.9.0
+  * Other dependencies updated
   * Minor improvements in mock handling
   * Minor code improvements on browserFactory and Modules